moved uuid out of sadfs::common
