+++ example/uuid.cpp
@@ -1,4 +1,4 @@
-#include <sadfs/common/uuid.hpp>
+#include <sadfs/uuid.hpp>
 
 #include <functional>
 #include <iostream>
@@ -9,16 +9,16 @@ int
 main()
 {
 	using namespace sadfs;
-	auto uuid = common::uuid::generate();
-	auto uuid_str = to_string(uuid);
+	auto id = uuid::generate();
+	auto id_str = to_string(id);
 
-	std::cout << "generated: " << uuid_str << "\n";
+	std::cout << "generated: " << id_str << "\n";
 
 	std::cout << "\n";
-	std::unordered_map<common::uuid, std::string> map{};
-	map.emplace(common::uuid::generate(), "generated first");
-	map.emplace(common::uuid::from_string(uuid_str), "parsed");
-	map.emplace(common::uuid::generate(), "generated again");
+	std::unordered_map<uuid, std::string> map{};
+	map.emplace(uuid::generate(), "generated first");
+	map.emplace(uuid::from_string(id_str), "parsed");
+	map.emplace(uuid::generate(), "generated again");
 
 	std::cout << "printing std::unordered_map of uuids\n"
 	          << "==================================================\n";
@@ -32,13 +32,13 @@ main()
 	// simulate (de)serialization to/from protobufs
 	std::cout << "\n";
 	auto bytes = std::string{};
-	uuid.serialize(std::back_inserter(bytes));
-	std::cout << uuid_str << " serialized to: \n";
+	id.serialize(std::back_inserter(bytes));
+	std::cout << id_str << " serialized to: \n";
 	for (auto byte : bytes)
 	{ std::cout << static_cast<int>(byte) << "\n"; }
 	std::cout << "\n";
 
-	auto new_uuid = common::uuid{};
-	new_uuid.deserialize(bytes.begin());
-	std::cout << "deserialized back to: " << to_string(new_uuid) << "\n";
+	auto new_id = uuid{};
+	new_id.deserialize(bytes.begin());
+	std::cout << "deserialized back to: " << to_string(new_id) << "\n";
 }
